Fix the circular dependency issue
