+++ x-pack/plugins/index_management/server/routes/api/inference_models/register_get_route.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { InferenceAPIConfigResponse } from '@kbn/ml-plugin/common/types/trained_models';
+import { InferenceAPIConfigResponse } from '@kbn/ml-trained-models-utils';
 import { addBasePath } from '..';
 import { RouteDependencies } from '../../../types';
