+++ x-pack/plugins/ml/common/types/trained_models.ts
@@ -12,6 +12,7 @@ import type {
   TotalFeatureImportance,
 } from '@kbn/ml-data-frame-analytics-utils';
 import type { IndexName, IndicesIndexState } from '@elastic/elasticsearch/lib/api/types';
+import type { InferenceAPIConfigResponse } from '@kbn/ml-trained-models-utils';
 import type { XOR } from './common';
 import type { MlSavedObjectType } from './saved_objects';
 
@@ -127,40 +128,6 @@ export interface PipelineDefinition {
   processors?: Array<Record<string, any>>;
   description?: string;
 }
-
-export type InferenceServiceSettings =
-  | {
-      service: 'elser';
-      service_settings: {
-        num_allocations: number;
-        num_threads: number;
-        model_id: string;
-      };
-    }
-  | {
-      service: 'openai';
-      service_settings: {
-        api_key: string;
-        organization_id: string;
-        url: string;
-      };
-    }
-  | {
-      service: 'hugging_face';
-      service_settings: {
-        api_key: string;
-        url: string;
-      };
-    };
-
-export type InferenceAPIConfigResponse = {
-  // Refers to a deployment id
-  model_id: string;
-  task_type: 'sparse_embedding' | 'text_embedding';
-  task_settings: {
-    model?: string;
-  };
-} & InferenceServiceSettings;
 
 export interface ModelPipelines {
   model_id: string;
